+++ intercom_flutter_web/example/integration_test/intercom_flutter_web_test.dart
@@ -61,6 +61,10 @@ void main() {
       });
     });
 
+    testWidgets('displayMessages', (WidgetTester _) async {
+      expect(plugin.displayMessages(), completes);
+    });
+
     testWidgets('displayMessageComposer', (WidgetTester _) async {
       expect(plugin.displayMessageComposer("message"), completes);
     });