+++ intercom_flutter_platform_interface/CHANGELOG.md
@@ -1,7 +1,17 @@
 # Changelog
 
+
 # 1.3.0
+
 - Updated `company` parameter inside `updateUser` method.
+
+# 1.2.2
+
+- Added method `displayMessages`.
+
+# 1.2.1
+
+- Updated dependencies
 
 # 1.2.0