+++ intercom_flutter_web/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Changelog
 
 # 0.3.0
+- 
 - Fixed `company` attribute mapping.
 - Updated dependency `intercom_flutter_platform_interface: ^1.3.0`
+
+# 0.2.2
+
+- Added method `displayMessages`.
+
+# 0.2.1
+
+- Updated dependencies
 
 # 0.2.0